+++ changelog.md
@@ -11,6 +11,7 @@
 - Possible to tell `eT_launch.py` to gracefully terminate eT using signals (`--signals`) or timeout (`--timeout`). eT-program/eT!763
 - Z-matrix tool can convert between xyz coordinates and Z-matrix. eT-program/eT!619
 - Davidson excited state solver can now handle cases where a requested root forms a complex pair with a higher root before convergence is reached (i.e., a false-positive complex pair); we thus avoid the commonly encountered "add one more root" error. eT-program/eT!772
+- Biorthonormalization works for a set of close states as well. eT-program/eT!775
 
 ### Structure
 - The test scripts now uses pathlib instead of os. eT-program/eT!756