+++ changelog.md
@@ -21,10 +21,13 @@
 - Cholesky orbital tool now handles decomposition of densities to get orbitals. eT-program/eT!762
 
 
+# eT v1.2.2
+### Bugfixes
+- Testing that the number of electrons is even for restricted closed calculations. See merge request eT-program/eT!777
+
 # eT v1.2.1
 ### Bugfixes
 - Reduced threshold in `uhf_energy_cs` to 1.0d-10 due to instabilities. See merge request eT-program/eT!766
-
 
 # eT v1.2
 ### Bugfixes